Modify createGrid function and use flexbox to form grid
createGrid was not implemented properly the first time, and so did not
form a 16x16 grid. Modified it so that a "row" div is created first and
each "square" div is appended to it. This created 16 "rows" with 16 "squares"/
"columns", which could be arranged into a 16x16 grid using flexbox.