Check for invalid inputs in click event listener and setGridSize
Before, if the user clicked cancel/returned an empty string, there would
be no grid generated. Default parameters were also added to createGrid
such that if the user only enters one number, then this would be the
number of rows and the number of columns would be set to equal it.
If no values are passed, then the default 16x16 grid is generated.
getGridSize is now setGridSize.